change main link in comment about viewbox
new link is more helpful, and one I looked at when trying to fix it, but didn't use as much in the end

first link in the second line is a stackoverflow response when I fixed it the second time, and was trying to find a reference to why I needed a view box, since I had gotten the idea for the strategy earlier, but had problems with it at the time, and reverted it.

second link in that comment is the link I included when trying to fix it the first time, based on my edit history. It is one of the links I looked at when trying to fix the scaling